refactor: update geo search input
